Move DAO holder
