+++ server/controllers/crud.controller.js
@@ -7,10 +7,6 @@ class CRUDController {
     try {
       const id = await this.dao.create(resource);
 
-      if (!id) {
-        throw new Error('500'); // TODO: 500?
-      }
-
       await onload(id);
       res.json({ id });
     } catch (err) {
