+++ server/controllers/crud/candidates.controller.js
@@ -1,5 +1,5 @@
 const CRUDController = require('../crud.controller');
-const db = require('../../dao');
+const db = require('../../dao/dao');
 const service = require('../../services/candidates.service');
 
 class CandidatesController extends CRUDController {
