+++ server/controllers/crud/feedbacks.controller.js
@@ -1,6 +1,6 @@
 const CRUDController = require('../crud.controller');
 
-const db = require('../../dao');
+const db = require('../../dao/dao');
 
 class FeedbacksController extends CRUDController {
   constructor() {
