+++ server/controllers/crud/hirings.controller.js
@@ -1,6 +1,6 @@
 const CRUDController = require('../crud.controller');
 
-const db = require('../../dao');
+const db = require('../../dao/dao');
 const service = require('../../services/hirings.service');
 
 class HiringsController extends CRUDController {
