+++ server/dao/dao.js
@@ -1,12 +1,12 @@
 const Bluebird = require('bluebird');
-const connection = Bluebird.promisifyAll(require('./dao/connection/connect'));
+const connection = Bluebird.promisifyAll(require('./connection/connect'));
 
-const Users = require('./dao/users.dao');
-const Candidates = require('./dao/candidates.dao');
-const Interviews = require('./dao/interviews.dao');
-const Hirings = require('./dao/hirings.dao');
-const Feedbacks = require('./dao/feedbacks.dao');
-const Vacancies = require('./dao/vacancies.dao');
+const Users = require('./users.dao');
+const Candidates = require('./candidates.dao');
+const Interviews = require('./interviews.dao');
+const Hirings = require('./hirings.dao');
+const Feedbacks = require('./feedbacks.dao');
+const Vacancies = require('./vacancies.dao');
 
 const users = new Users(connection);
 const candidates = new Candidates(connection);
